# swagger.yaml
openapi: 3.0.0
info:
  title: delta_capita_v2
  version: 1.0.0
  description: API documentation generated from Postman collection

servers:
  - url: http://localhost:3000

paths:
  /:
    get:
      summary: Base welcome endpoint
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                basket:
                  type: array
                  items:
                    type: string
              example:
                basket: ["Apple", "Apple", "Banana", "Melon", "Melon", "Lime", "Lime", "Lime"]
      responses:
        '200':
          description: Welcome message
          content:
            application/json:
              schema:
                type: object
                properties:
                  welcome:
                    type: string

  /api/products:
    get:
      summary: List all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    price:
                      type: number

    post:
      summary: Create or update a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
              properties:
                name:
                  type: string
                price:
                  type: number
      responses:
        '201':
          description: Product added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number

  /api/products/{productName}:
    delete:
      summary: Delete a product by name
      parameters:
        - in: path
          name: productName
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /api/offers:
    post:
      summary: Create an offer for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productName
                - type
                - description
              properties:
                productName:
                  type: string
                type:
                  type: string
                  enum: [2FOR1, 3FOR2]
                description:
                  type: string
              example:
                productName: Melon
                type: 2FOR1
                description: Buy one get one free
      responses:
        '201':
          description: Offer added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string